```text chore: Refactor getNextIndex function to fix off-by-one error
Fix the off-by-one error in the getNextIndex function of scripts.js. The condition for checking if the next index is greater than or equal to the length of formState should be changed to just greater than. This change ensures that the correct index is returned when determining the next form element.